requirements.txt
albumentations==1.3.1
matplotlib==3.8.0
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==10.0.1
Pillow==12.0.0
scikit_learn==1.3.1
seaborn==0.13.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1